+++ packages/plugin-gallery/web/src/components/gallery-controls/gallery-items-sortable.js
@@ -487,7 +487,7 @@ export class SortableComponent extends Component {
   propsToState(props) {
     return {
       items: props.items,
-      editedImage: props?.items?.[this.state?.editedImageIndex],
+      editedItem: props?.items?.[this.state?.editedItemIndex],
     };
   }